+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add internal iterative reductions (IIR) (9.66 +- 5.53)
 * Use threats in quiet move history (9.28 +- 5.28 at LTC with threat calculation, 18.34 +- 7.98 STC excluding threat calculation)
 * Add late move pruning (9.63 +- 5.59)
+* Add double extensions (9.28 +- 5.26)
 * Revert net scale change with a view to using WDL normalisation instead (9.08 +- 5.12)
 * Avoid re-allocating thread data before each search (8.95 +- 5.20 with 2 threads vs 2 threads, +~20 Elo in 2v1 thread odds)
 * Add 1-ply continuation history (8.88 +- 5.20)